requirements.txt
gradio==4.44.1
pillow==10.4.0
numpy==2.1.1
audioop-lts==0.2.1
opencv-python==4.10.0.84
